backend/budgetTracker/requirements.txt
asgiref==3.10.0
django==5.2.7
django-cors-headers==4.9.0
django_filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
python-decouple==3.8
sqlparse==0.5.3
